ci: KEEP-325 add events workspace CI coverage and PR deploy label
- New pr-checks-events.yml runs lint, typecheck, unit, and integration
  tests on PRs touching keeperhub-events/**. Integration job brings up
  only test-anvil/test-localstack/test-redis from the existing test
  compose profile (skips the heavy main-app peers).
- Rename event-tracker tests/e2e/ to tests/integration/ to match the
  main app's tier convention; the tests are integration tests by
  function (in-process tracker against real anvil/redis/localstack),
  not against a built/deployed artifact. Adds test:unit and
  test:integration scripts.
- New deploy-pr-events label: when present alongside
  deploy-pr-environment, builds a PR-specific event-tracker image and
  deploys the events satellite from it. Without the label, the PR env
  continues to use the staging event-latest tag (current behavior).
  Build pushes only the sha-tagged tag, never event-latest, so
  concurrent staging deploys are unaffected.
