+++ .github/workflows/deploy-pr-environment.yaml
@@ -12,6 +12,7 @@ jobs:
     outputs:
       should-deploy: ${{ steps.check.outputs.should-deploy }}
       should-test: ${{ steps.check.outputs.should-test }}
+      should-deploy-events: ${{ steps.check.outputs.should-deploy-events }}
     steps:
       - name: Checkout code (for change detection)
         if: github.event.action == 'synchronize'
@@ -24,9 +25,15 @@ jobs:
         run: |
           echo "should-deploy=false" >> $GITHUB_OUTPUT
           echo "should-test=false" >> $GITHUB_OUTPUT
+          echo "should-deploy-events=false" >> $GITHUB_OUTPUT
 
           HAS_DEPLOY_LABEL="${{ contains(github.event.pull_request.labels.*.name, 'deploy-pr-environment') }}"
           HAS_TEST_LABEL="${{ contains(github.event.pull_request.labels.*.name, 'run-e2e-tests-pr-deploy') }}"
+          # Events PR build is opt-in alongside the standard PR deploy. Both
+          # labels must be present: events satellite ships into the PR
+          # namespace created by deploy-pr-environment, so it has no meaning
+          # without that label. See KEEP-325.
+          HAS_EVENTS_LABEL="${{ contains(github.event.pull_request.labels.*.name, 'deploy-pr-events') }}"
           ACTION="${{ github.event.action }}"
           ADDED_LABEL="${{ github.event.label.name }}"
 
@@ -36,8 +43,22 @@ jobs:
               if [[ "$HAS_TEST_LABEL" == "true" ]]; then
                 echo "should-test=true" >> $GITHUB_OUTPUT
               fi
+              if [[ "$HAS_EVENTS_LABEL" == "true" ]]; then
+                echo "should-deploy-events=true" >> $GITHUB_OUTPUT
+              fi
             elif [[ "$ADDED_LABEL" == "run-e2e-tests-pr-deploy" ]]; then
               echo "should-test=true" >> $GITHUB_OUTPUT
+            elif [[ "$ADDED_LABEL" == "deploy-pr-events" ]]; then
+              # Events-only re-trigger: only meaningful if deploy-pr-environment
+              # is already present. Re-run full deploy so the events satellite
+              # picks up the freshly-built PR events image.
+              if [[ "$HAS_DEPLOY_LABEL" == "true" ]]; then
+                echo "should-deploy=true" >> $GITHUB_OUTPUT
+                echo "should-deploy-events=true" >> $GITHUB_OUTPUT
+                if [[ "$HAS_TEST_LABEL" == "true" ]]; then
+                  echo "should-test=true" >> $GITHUB_OUTPUT
+                fi
+              fi
             fi
           elif [[ "$ACTION" == "synchronize" || "$ACTION" == "opened" || "$ACTION" == "reopened" ]]; then
             NEEDS_DEPLOY=true
@@ -56,6 +77,9 @@ jobs:
 
             if [[ "$HAS_DEPLOY_LABEL" == "true" && "$NEEDS_DEPLOY" == "true" ]]; then
               echo "should-deploy=true" >> $GITHUB_OUTPUT
+              if [[ "$HAS_EVENTS_LABEL" == "true" ]]; then
+                echo "should-deploy-events=true" >> $GITHUB_OUTPUT
+              fi
             fi
             # Run tests if both labels present (even without deploy)
             if [[ "$HAS_DEPLOY_LABEL" == "true" && "$HAS_TEST_LABEL" == "true" ]]; then
@@ -140,9 +164,81 @@ jobs:
             type=registry,ref=${{ steps.login-ecr.outputs.registry }}/${{ env.AWS_ECR_NAME }}:${{ matrix.image.cache_key }}
           cache-to: type=registry,ref=${{ steps.login-ecr.outputs.registry }}/${{ env.AWS_ECR_NAME }}:${{ matrix.image.cache_key }},mode=max
 
+  # Builds a PR-specific event-tracker image when the deploy-pr-events
+  # label is present alongside deploy-pr-environment. Without it, the PR
+  # environment uses the staging events image (event-latest tag). Pushed
+  # to the staging events ECR with an event-${sha_short} tag; never pushes
+  # event-latest so concurrent staging deploys are not affected. See KEEP-325.
+  build-events-image:
+    needs: check-labels
+    if: needs.check-labels.outputs.should-deploy-events == 'true'
+    runs-on: ubuntu-latest
+    environment: staging
+    env:
+      REGION: ${{ vars.TO_REGION }}
+      EVENTS_ECR_NAME: keeperhub-event-staging
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v6
+        with:
+          ref: ${{ github.event.pull_request.head.sha }}
+
+      - name: Configure AWS credentials
+        uses: aws-actions/configure-aws-credentials@v6
+        with:
+          aws-access-key-id: ${{ secrets.TO_AWS_ACCESS_KEY_ID }}
+          aws-secret-access-key: ${{ secrets.TO_AWS_SECRET_ACCESS_KEY }}
+          aws-region: ${{ env.REGION }}
+
+      - name: Login to AWS ECR
+        id: login-ecr
+        uses: aws-actions/amazon-ecr-login@v2
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v4
+
+      - name: Extract commit hash
+        id: vars
+        run: echo "sha_short=$(git rev-parse --short HEAD)" >> "$GITHUB_OUTPUT"
+
+      - name: Check if events image exists
+        id: check-image
+        run: |
+          IMAGE_TAG="${{ steps.vars.outputs.sha_short }}"
+          if aws ecr describe-images \
+            --repository-name ${{ env.EVENTS_ECR_NAME }} \
+            --image-ids imageTag=event-${IMAGE_TAG} \
+            --region ${{ env.REGION }} >/dev/null 2>&1; then
+            echo "exists=true" >> $GITHUB_OUTPUT
+          else
+            echo "exists=false" >> $GITHUB_OUTPUT
+          fi
+
+      - name: Build and push events image
+        if: steps.check-image.outputs.exists != 'true'
+        uses: docker/build-push-action@v7
+        with:
+          context: ./keeperhub-events
+          file: ./keeperhub-events/event-tracker/Dockerfile
+          push: true
+          provenance: false
+          tags: |
+            ${{ steps.login-ecr.outputs.registry }}/${{ env.EVENTS_ECR_NAME }}:event-${{ steps.vars.outputs.sha_short }}
+          cache-from: |
+            type=registry,ref=${{ steps.login-ecr.outputs.registry }}/${{ env.EVENTS_ECR_NAME }}:cache
+          cache-to: type=registry,ref=${{ steps.login-ecr.outputs.registry }}/${{ env.EVENTS_ECR_NAME }}:cache,mode=max
+
   deploy:
-    needs: [check-labels, build-images]
-    if: needs.check-labels.outputs.should-deploy == 'true'
+    needs: [check-labels, build-images, build-events-image]
+    # Run when the deploy label is set, build-images succeeded, and the
+    # optional events build either succeeded or was skipped (label absent).
+    # !cancelled() prevents the default "needed job skipped -> skip me" rule
+    # from blocking deploy when build-events-image is intentionally skipped.
+    if: |
+      !cancelled() &&
+      needs.check-labels.outputs.should-deploy == 'true' &&
+      needs.build-images.result == 'success' &&
+      (needs.build-events-image.result == 'success' || needs.build-events-image.result == 'skipped')
     runs-on: ubuntu-latest
     environment: staging
     env:
@@ -276,6 +372,10 @@ jobs:
           DB_PASSWORD: ${{ steps.db-password.outputs.password }}
           SERVICE_ACCOUNT_ROLE_ARN: ${{ vars.TO_SERVICE_ACCOUNT_ROLE_ARN }}
           TEST_PARA_USER_SHARE: ${{ secrets.TEST_PARA_USER_SHARE }}
+          # PR-specific events image tag if deploy-pr-events label was set
+          # (build-events-image job ran). Otherwise stays on the staging
+          # event-latest tag.
+          EVENTS_IMAGE_TAG: ${{ needs.check-labels.outputs.should-deploy-events == 'true' && format('event-{0}', steps.vars.outputs.sha_short) || 'event-latest' }}
         run: |
           # Percent-encode DB_PASSWORD so base64 characters (+/=) don't break URL parsing
           # in the pod. Init containers previously encoded at runtime but that path was
