+++ keeperhub-events/event-tracker/package.json
@@ -11,7 +11,8 @@
     "lint:fix": "biome check --write .",
     "test": "vitest run",
     "test:watch": "vitest",
-    "test:e2e": "vitest run"
+    "test:unit": "vitest run tests/unit",
+    "test:integration": "vitest run tests/integration"
   },
   "dependencies": {
     "@aws-sdk/client-sqs": "^3.1005.0",
